tokens are long

+++ src/schedproc.h
@@ -42,7 +42,7 @@ EXTERN struct schedproc {
 	bitchunk_t cpu_mask[BITMAP_CHUNKS(CONFIG_MAX_CPUS)]; /* what CPUs is the process allowed to run on */
 
     /* Tokens */
-    int tokens;        /* number of tokens */
+    long tokens;        /* number of tokens */
     clock_t tokens_updated; /* last sys time value when updated tokens for this process */
 } schedproc[NR_PROCS];
